Project05
This was a project for my Analysis of Algorithms course at UNAM, i
rescued it from the bin. It´s not complete since i lost the final
version which actually i was evaluated.

This project was about implementing Dijkstra´s shortest path algorithm
in an graph represented as an XML. It can be tested with Apache Ant.